Moved SkiRuns to locationManager package
